requirements.txt
airtest==1.1.8
airtest-selenium==1.0.3
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
delayed-assert==0.3.2
dill==0.3.2
hrpc==1.0.8
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
Jinja2==2.11.2
loguru==0.5.2
MarkupSafe==1.1.1
more-itertools==8.5.0
mss==4.0.3
opencv-contrib-python==3.4.8.29
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
pocoui==1.0.79
py==1.9.0
pynput==1.7.1
#pyobjc-core==6.2.2
#pyobjc-framework-Cocoa==6.2.2
#pyobjc-framework-Quartz==6.2.2
pyparsing==2.4.7
pytest==5.3.5
pytest-reportportal==1.0.9
python-xlib==0.27
pywinauto==0.6.3
PyYAML==5.3.1
reportportal-client==3.2.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wcwidth==0.2.5
websocket-client==0.48.0
zipp==3.1.0
numpy==1.19.2

